@@ -6,5 +6,5 @@ public class Constants {
   public static final String ACL_REL_CONFIG_REDIS_KEY = "rel_configs";
   public static final String SIGNOUT_REDIS_KEY = "SIGNOUT_%s";
   public static final String NEED_REFRESH_MAPPING_CACHE_MARKER_KEY =
-          "NEED_REFRESH_MAPPING_CACHE_MARKER";
+      "NEED_REFRESH_MAPPING_CACHE_MARKER";
 }
