@@ -11,21 +11,22 @@ import redis.clients.jedis.JedisPoolConfig;
 
 @Configuration
 public class RedisConfig {
+
   @Bean
   @ConditionalOnProperty(value = "redis.enabled", havingValue = "true")
   public JedisPool jedisPool(
       @Value("${redis.hostname}") String redisHost,
       @Value("${redis.port}") Integer redisPort,
       @Value("${redis.ssl}") Boolean sslEnabled,
-      @Value("${redis.password}") String redisPassword,
-      JedisPoolConfig poolConfig) {
+      @Value("${redis.password}") String redisPassword) {
 
     if (Strings.isEmpty(redisPassword)) redisPassword = null;
-    return new JedisPool(poolConfig, redisHost, redisPort, 2000, redisPassword, sslEnabled);
+    JedisPool pool =
+        new JedisPool(buildPoolConfig(), redisHost, redisPort, 2000, redisPassword, sslEnabled);
+    return pool;
   }
 
-  @Bean
-  public JedisPoolConfig buildPoolConfig() {
+  private JedisPoolConfig buildPoolConfig() {
     final JedisPoolConfig poolConfig = new JedisPoolConfig();
     poolConfig.setMaxTotal(128);
     poolConfig.setMaxIdle(128);
@@ -33,10 +34,11 @@ public class RedisConfig {
     poolConfig.setTestOnBorrow(true);
     poolConfig.setTestOnReturn(true);
     poolConfig.setTestWhileIdle(true);
-    poolConfig.setMinEvictableIdleTimeMillis(Duration.ofSeconds(60).toMillis());
-    poolConfig.setTimeBetweenEvictionRunsMillis(Duration.ofSeconds(30).toMillis());
+    poolConfig.setMinEvictableIdleTime(Duration.ofSeconds(60));
+    poolConfig.setTimeBetweenEvictionRuns(Duration.ofSeconds(30));
     poolConfig.setNumTestsPerEvictionRun(3);
     poolConfig.setBlockWhenExhausted(true);
+    poolConfig.setJmxEnabled(false); // this is ok, spring will register it
     return poolConfig;
   }
 }
