@@ -4,10 +4,10 @@ import authserver.acl.Acl;
 import authserver.common.AclOperation;
 import authserver.common.AclOperationDto;
 import com.google.common.base.Stopwatch;
+import jakarta.validation.Valid;
 import java.util.HashSet;
 import java.util.Set;
 import java.util.concurrent.TimeUnit;
-import javax.validation.Valid;
 import lombok.extern.slf4j.Slf4j;
 import org.example.authserver.config.AppProperties;
 import org.example.authserver.entity.AclsRequestDTO;
