@@ -1,8 +1,8 @@
 package org.example.authserver.entity;
 
 import authserver.acl.Acl;
+import jakarta.validation.constraints.NotNull;
 import java.util.List;
-import javax.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
