@@ -1,18 +1,16 @@
 package org.example.authserver.entity;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
-import com.vladmihalcea.hibernate.type.json.JsonBinaryType;
+import jakarta.persistence.*;
 import java.io.Serializable;
 import java.util.List;
 import java.util.Objects;
-import javax.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import org.hibernate.annotations.Type;
-import org.hibernate.annotations.TypeDef;
-import org.springframework.http.HttpMethod;
+import org.hibernate.annotations.JdbcTypeCode;
+import org.hibernate.type.SqlTypes;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @Builder
@@ -20,7 +18,6 @@ import org.springframework.http.HttpMethod;
 @NoArgsConstructor
 @AllArgsConstructor
 @Entity(name = "mappings")
-@TypeDef(name = "jsonb", typeClass = JsonBinaryType.class)
 public class MappingEntity implements Serializable {
 
   @Id private String id;
@@ -30,15 +27,15 @@ public class MappingEntity implements Serializable {
 
   private String path;
 
-  @Type(type = "jsonb")
+  @JdbcTypeCode(SqlTypes.JSON)
   @Column(columnDefinition = "jsonb")
   private List<HeaderMappingKey> headerMapping;
 
-  @Type(type = "jsonb")
+  @JdbcTypeCode(SqlTypes.JSON)
   @Column(columnDefinition = "jsonb")
   private BodyMapping bodyMapping;
 
-  @Type(type = "jsonb")
+  @JdbcTypeCode(SqlTypes.JSON)
   @Column(columnDefinition = "jsonb")
   private List<String> roles;
 
