@@ -3,21 +3,13 @@ package org.example.authserver.service;
 import authserver.acl.Acl;
 import authserver.acl.AclRelationConfig;
 import io.micrometer.core.annotation.Timed;
-import java.nio.charset.StandardCharsets;
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Map;
-import java.util.Optional;
-import java.util.Set;
+import java.util.*;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.Executor;
 import java.util.concurrent.Executors;
 import java.util.stream.Collectors;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.commons.codec.digest.MurmurHash3;
 import org.example.authserver.Utils;
 import org.example.authserver.entity.RelCache;
 import org.example.authserver.repo.RelCacheRepository;
@@ -106,8 +98,7 @@ public class CacheServiceImpl implements CacheService {
     for (String tag : relations) {
       Set<String> nestedTags = new HashSet<>(relations);
       nestedTags.remove(tag); // remove current
-      String compositeIdKey = getCompositeIdKey(principal, tag, path, revision);
-      String id = getKeyHash(compositeIdKey);
+      String id = UUID.randomUUID().toString();
       Acl parsedTag = Utils.parseTag(tag);
       if (parsedTag == null) {
         continue;
@@ -133,18 +124,5 @@ public class CacheServiceImpl implements CacheService {
       // lock.
       log.debug("Cache saving problems", e);
     }
-  }
-
-  private String getKeyHash(String compositeIdKey) {
-    long[] vec = MurmurHash3.hash128(compositeIdKey.getBytes(StandardCharsets.UTF_8));
-    StringBuilder sb = new StringBuilder();
-    for (long l : vec) {
-      sb.append(Long.toHexString(l));
-    }
-    return sb.toString();
-  }
-
-  private String getCompositeIdKey(String principal, String tag, String path, Long revision) {
-    return String.format("%s+%s+%s+%s", principal, tag, path, revision);
   }
 }
